Added Go templating function processing, such as reReplaceAll, match, and humanize.
Feature: Added Go templating function processing, such as reReplaceAll, match, and humanize.
Update: Because the base of the templating is done through an existing Prometheus repository, the go.mod and go.sum files have been re-created. This includes running the "go mod tidy".

A large chunk of the functions listed under Grafana functions (https://grafana.com/docs/grafana/latest/alerting/fundamentals/annotation-label/template-functions) is usable with the bridge.

From what I could tell, some template functions need more data, which does not pass in the JSON. These functions will log output explaining the function is not supported by the bridge. The functionality for these excluded functions still exists and would be simple to enable if the data becomes avaliable. Also, a few of the template functions in the template.go file were not listed in the Grafana documentation, so those are also excluded.